more forgiving api call matching
List all commands, see if the provided command is in the list. if it's
not, see if it is a substring of an item in that list. If so, try the
command. If the command errors then that's probably not the command we
wanted to run, so let's try again with the next contender.